requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.8
django-debug-toolbar==4.2.0
django-grappelli==3.0.8
djangorestframework==3.14.0
idna==3.6
pillow==10.2.0
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
